refactor: добавлены index.ts к компонентам
