+++ packages/ui/src/components/Button/Button.mdx
@@ -7,8 +7,8 @@ route: /components/Button
 import { Playground, Props } from 'docz'
 
 import { ThemeSelect, TypoSelect } from '../../helpers/ThemeSelect'
+import { IconDownload } from '../Icon'
 import { ThemeBackground } from '../../helpers/ThemeBackground'
-import { IconDownload } from '../Icon/Icons/IconDownload'
 
 import { Button } from './Button'
 import { ActionButton } from './ActionButton'
