+++ packages/ui/src/components/Card/Card.mdx
@@ -10,11 +10,9 @@ import { ThemeSelect } from '../../helpers/ThemeSelect'
 import { ThemeBackground } from '../../helpers/ThemeBackground'
 import { IconPlaceholder } from '../../helpers/IconPlaceholder'
 import { ImagePlaceholder } from '../../helpers/ImagePlaceholder'
-import { Button } from '../Button/Button'
-import { Cell, CellIcon } from '../Cell/Cell'
-import { CellListItem } from '../Cell/CellListItem'
-import { Disclosure } from '../Cell/CellDisclosure'
-import { TextBox, Title, SubTitle, BiggerTitle, BigTitle } from '../TextBox/TextBox'
+import { Button } from '../Button'
+import { Cell, CellIcon, CellListItem, Disclosure } from '../Cell'
+import { TextBox, Title, SubTitle, BiggerTitle, BigTitle } from '../TextBox'
 
 import { Card } from './Card'
 import { CardBody } from './CardBody'
