+++ packages/ui/src/components/Cell/Cell.mdx
@@ -7,7 +7,7 @@ route: /components/Cell
 import { Playground, Props } from 'docz';
 
 import { Cell, CellIcon } from './Cell';
-import { TextBox } from '../TextBox/TextBox';
+import { TextBox } from '../TextBox';
 
 import { ThemeSelect } from '../../helpers/ThemeSelect'
 import { ThemeBackground } from '../../helpers/ThemeBackground'
