+++ packages/ui/src/components/Grid/index.ts
@@ -1,0 +1,1 @@
+export * from './Grid';
