+++ packages/ui/src/components/Header/Header.mdx
@@ -8,7 +8,7 @@ import { Playground, Props } from 'docz';
 
 import Logo from "../../../.storybook/public/images/logo.png";
 import { Header, HeaderBack, HeaderLogo, HeaderSubTitle, HeaderTitle, HeaderContent, HeaderButton } from './Header';
-import { Cart } from '../Cart/Cart';
+import { Cart } from '../Cart';
 import { ThemeSelect } from '../../helpers/ThemeSelect'
 import { ThemeBackground } from '../../helpers/ThemeBackground'
