+++ packages/ui/src/components/Stepper/index.ts
@@ -1,0 +1,1 @@
+export * from './Stepper';
