+++ packages/ui/src/components/TextField/TextField.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useState, useCallback } from 'react';
 import styled, { css } from 'styled-components';
 import { colors, scalingPixelBasis } from '@sberdevices/plasma-tokens';
 
-import { StyledInputRoot, StyledInput, InputProps } from '../Input/Input';
+import { StyledInputRoot, StyledInput, InputProps } from '../Input';
 
 // В этих константах задаем размеры в em, чтобы не зависеть напрямую от пикселей
 // В то же время в числителях - значения в пикселях, взятые из макета
